Moved the DTO classes to respective sub-packages
