+++ src/main/java/in/gagan/base/framework/controller/UserController.java
@@ -41,8 +41,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import in.gagan.base.framework.dto.UpdateUserDTO;
-import in.gagan.base.framework.dto.UserDTO;
+import in.gagan.base.framework.dto.user.UpdateUserDTO;
+import in.gagan.base.framework.dto.user.UserDTO;
 import in.gagan.base.framework.exception.UsernameExistException;
 import in.gagan.base.framework.service.user.UserRegisterationService;
 import in.gagan.base.framework.util.UserHelperUtil;
