+++ src/main/java/in/gagan/base/framework/dto/user/UsernamePasswordAuthDTO.java
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package in.gagan.base.framework.dto;
+package in.gagan.base.framework.dto.user;
 
 /**
  * This DTO captures the username and password from the user for authentication by the spring security.
