+++ src/main/java/in/gagan/base/framework/filter/JWTUsernamePasswordAuthFilter.java
@@ -43,7 +43,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import in.gagan.base.framework.component.JWTProps;
 import in.gagan.base.framework.constant.ApplicationConstants;
-import in.gagan.base.framework.dto.UsernamePasswordAuthDTO;
+import in.gagan.base.framework.dto.user.UsernamePasswordAuthDTO;
 import in.gagan.base.framework.service.user.JWTService;
 
 /**
