+++ src/main/java/in/gagan/base/framework/service/user/CustomUserDetailsService.java
@@ -34,8 +34,8 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import in.gagan.base.framework.dao.user.UserDAO;
-import in.gagan.base.framework.dto.UserDetailsDTO;
-import in.gagan.base.framework.dto.UserRoleDTO;
+import in.gagan.base.framework.dto.user.UserDetailsDTO;
+import in.gagan.base.framework.dto.user.UserRoleDTO;
 import in.gagan.base.framework.util.ExceptionHelperUtil;
 
 /**
