+++ src/main/java/in/gagan/base/framework/service/user/UserRegisterationService.java
@@ -19,8 +19,8 @@
 
 package in.gagan.base.framework.service.user;
 
-import in.gagan.base.framework.dto.UpdateUserDTO;
-import in.gagan.base.framework.dto.UserDTO;
+import in.gagan.base.framework.dto.user.UpdateUserDTO;
+import in.gagan.base.framework.dto.user.UserDTO;
 import in.gagan.base.framework.entity.User;
 import in.gagan.base.framework.exception.UsernameExistException;
