+++ src/main/java/in/gagan/base/framework/service/user/UserRegisterationServiceImpl.java
@@ -27,8 +27,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import in.gagan.base.framework.constant.ApplicationConstants;
-import in.gagan.base.framework.dto.UpdateUserDTO;
-import in.gagan.base.framework.dto.UserDTO;
+import in.gagan.base.framework.dto.user.UpdateUserDTO;
+import in.gagan.base.framework.dto.user.UserDTO;
 import in.gagan.base.framework.entity.User;
 import in.gagan.base.framework.entity.VerificationToken;
 import in.gagan.base.framework.exception.UsernameExistException;
